requirements.txt
behave==1.2.6
behave-django==1.4.0
cairocffi==1.4.0
CairoSVG==2.5.2
certifi==2022.9.24
cffi==1.15.1
chardet==5.0.0
cssselect==1.2.0
cssselect2==0.7.0
defusedxml==0.7.1
Django==4.0.7
django-allauth==0.51.0
django-filter==22.1
djangorestframework==3.14.0
gunicorn==20.1.0
html5lib==1.1
idna==3.4
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
oauthlib==3.2.2
olefile==0.46
parse==1.19.0
parse-type==0.6.0
pdfrw==0.4
Pillow==9.3.0
pycparser==2.21
pyphen==0.12.0
python3-openid==3.2.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
tinycss==0.4
tinycss2==1.2.1
urllib3==1.26.13
weasyprint==57.1
webencodings==0.5.1
